pyproject.toml
[project]
name = "onepkg"
version = "0.6.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6",
    "cyclopts>=3.20.0",
    "rich>=13.0.0",
]
description = "One manager to manage them all - unified CLI for brew, cargo, uv, go, and more"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.8.0",
]

[project.scripts]
onepkg = "onepkg.cli:main"

[tool.setuptools.package-data]
onepkg = ["specs.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
